refactor(frontend): split main.js leaf modules + OTA flow + radio mirror updates
Phase A of the main.js refactor: the five low-risk leaf modules
(state, utils, localization, logos, api) are extracted into their
own files, leaving views in main.js for a later Phase B. Net diff
is -668/+991 because the new module files duplicate the size of
the removed sections plus added documentation. main.js itself
shrinks 3571 → 3015 lines (-555).

A series of bugs uncovered during local-build testing of the
refactor are folded in here rather than as follow-ups, per the
"no quick fix only" rule.

NEW MODULES

- state.js (106 lines): central state object + localStorage helpers
  (loadCachedBoxes / saveCachedBoxes / loadLastBox / saveLastBox).
  Both cache helpers now filter entries whose host is 203.0.113/24,
  127/8, or 0.0.0.0 / 255.255.255.255 — an older build of the app
  could persist the Bose USB-gadget TEST-NET-3 address into
  localStorage, after which every launch loaded the bad host and
  silently broke /api/box/volume + presets until discovery
  overwrote the cache 4 s later (and even longer if mDNS lagged).
- utils.js (123 lines): $, escapeHtml, escapeAttr, decodeXmlEntities
  (now includes &#NN; and &#xHH; numeric refs), formatNumber,
  debounce, sleep, plus self-wiring modal helpers (confirmWarn,
  closeWarn, showError, showToast).
- localization.js (375 lines): COUNTRIES, ORDERS, GENRE_CORE,
  GENRE_BY_COUNTRY, COUNTRY_DE, SKIP_TAGS, GENRE_ALIAS, GENRE_DE,
  translateCountry, canonGenre, translateGenre, translateTags,
  flagFromCC. User-facing German strings deliberately left in place
  — Phase C (i18n) will move them into a translation bundle.
- logos.js (102 lines): extractHost, rootDomain, iconServicesFor,
  stationLogoCandidates, logoImgTag, bestLogoForStation,
  stationLogoChain. Google's favicon service removed from
  iconServicesFor on user request — only DuckDuckGo's
  privacy-friendly icon service is queried now.
- api.js (52 lines): re-exports all Wails-generated bindings and
  BrowserOpenURL so views don't need to know the wailsjs path,
  plus a small boxURL() helper for the endpoints not yet wrapped
  on the Go side.

main.js: imports from the new modules, all duplicated definitions
removed, all five leaf-module identifiers consumed via named
imports.

OTA + UPDATE-INDICATOR FIXES

- discovery package now broadcasts the agent's build stamp via the
  mDNS TXT record (build=YYYY-MM-DD-HHMM alongside the existing
  version key). Without this, the desktop app's tab badge and
  Musik-hören update banner could not distinguish two binaries
  that share a git-describe version but differ in build stamp —
  drift went silent in two of three update indicators while
  Box-Einstellungen flagged it correctly, producing the confusing
  "the box says it needs an update but the rest of the UI denies
  it" state.
  - discovery.Config gets a Build field
  - discovery.Instance gets a Build field
  - register() includes "build=" in the TXT record
  - Browse parses "build" out of incoming TXT entries
- cmd/agent/main.go passes buildStamp to discovery.Announce.
- desktop-app/app.go BoxInfo gets a Build field, populated from
  inst.Build by DiscoverBoxes.
- frontend updateSettingsTabBadge() now considers BOTH version
  and build for drift detection. It also treats "we have a build,
  box reports none" as drift, which catches older agents that do
  not yet broadcast the new build= TXT key — without that, a
  pre-this-release box would silently appear up to date.
- frontend checkBoxUpdate() (Musik-hören banner) compares both
  version and build; previously it short-circuited as soon as
  version matched, hiding the banner on stamp-only drift.
- frontend bootFromCache() now calls checkBoxUpdate() on startup
  when a box is restored from cache. Previously, the Musik-hören
  banner only fired when discoverBoxes() detected a "changed"
  diff (host/port/version/friendlyName), so an app booting against
  a cached box whose mDNS-version still matched would never run
  the agent-version comparison even if builds differed.
- frontend doBoxUpdate() now polls BoxAgentVersion() for up to
  3 minutes after upload, keeping both update buttons locked with
  a "Box läuft neu hoch (...)" status until the box answers with
  a build matching the app's appBuild — or the deadline fires.
  Previously the buttons re-enabled after a fixed 13 s timer,
  which invited a second click during the ~70 s TIME_WAIT-tolerant
  restart window and resulted in a corrupted second OTA upload.
- frontend update banner template gets a <br> after "Box Software
  Update verfügbar" so the longer build-aware label wraps cleanly
  instead of running inline on a single very wide line.

RADIO MIRROR + RETRY HARDENING

- internal/radiobrowser/radiobrowser.go: hardcoded Mirrors list
  trimmed to current reality. radio-browser.info's own discovery
  endpoint
  (https://all.api.radio-browser.info/json/servers) only returns
  de1.api.radio-browser.info as alive today; de2/at1/nl1/fi1 no
  longer resolve via DNS. New list:
    - https://de1.api.radio-browser.info/json
    - https://all.api.radio-browser.info/json   (DNS round-robin
      fallback, self-heals if the project re-adds mirrors)
    - https://91.98.4.78/json                   (de1 IPv4, last
      resort when DNS itself is down)
- fetchJSON now derives a per-attempt context with a 3 s timeout
  instead of sharing the outer handler ctx across all five
  mirrors. Without this, the first slow mirror used to consume
  the full 8 s budget and failover was effectively dead — the
  exact symptom that produced the "first try 502, second try
  works" report.
- handleRadioSearch, handleRadioTop, handleRadioTags, and
  handleRadioLanguages outer context bumped 8 s → 18 s. With
  three mirrors at 3 s each plus margin, the budget can actually
  accommodate the per-mirror retries.

PIPELINE / META

- desktop-app/frontend/wailsjs/go/models.ts regenerated to expose
  the new BoxInfo.Build field across the Wails bridge.
